refactor: SRP-split delegation.rs (576→10 files, all <50 eff lines)
Split session/delegation.rs into directory module following the same
pattern used for session/index/:

- mod.rs      (15 eff) — re-exports + test module declarations
- beta.rs     (34 eff) — BetaPosterior struct + math
- config.rs   (27 eff) — DelegationConfig + DEFAULT_* constants
- env.rs       (9 eff) — CODETETHER_DELEGATION_ENABLED override
- state.rs    (27 eff) — DelegationState struct + key()
- state_mut.rs (33 eff) — update, score, ensure, delegate_to
- state_query.rs (36 eff) — rank_candidates, shrink_cold_start
- tests_beta.rs   (20 eff) — 2 BetaPosterior tests
- tests_state.rs  (37 eff) — 3 mutation tests
- tests_query.rs  (39 eff) — 4 ranking/config tests

All 9 original tests preserved. Zero external API changes.
